feat(image): make cms/image component a general component-library component, move it from cms folder to main folder
