+++ frontend/packages/component-library/src/image/Image.tsx
