+++ frontend/packages/component-library/src/image/README.md
@@ -1,4 +1,4 @@
-# `component-library/cms/image`
+# `component-library/image`
 
 ## Consuming This Component
 
@@ -6,12 +6,10 @@ This package exports one styled React component: [Image](Image.tsx).
 You can import it like this:
 
 ```javascript
-import Image from '@code-dot-org/component-library/cms/image';
+import Image from '@code-dot-org/component-library/image';
 ```
 
-_Note:_ This component is used primarily for the Contentful CMS.
-
 For guidelines on how to use this component and the features it
-offers, [visit Storybook](https://code-dot-org.github.io/code-dot-org/component-library-storybook/?path=/docs/cms-image--docs).
+offers, [visit Storybook](https://code-dot-org.github.io/code-dot-org/component-library-storybook/?path=/docs/designsystem-image--docs).
 Or run storybook locally and go
 to [CMS / Image](http://localhost:6006/?path=/docs/designsystem-image--docs).
