+++ frontend/packages/component-library/src/image/__tests__/Image.test.tsx
