+++ frontend/packages/component-library/src/image/image.module.scss
