+++ frontend/packages/component-library/src/image/index.ts
